Depend on InverseFunctions
Also remove unused hard-dep on Functors
